Various Karma cleanup in docs and config files
I left the commented out WITH_KARMA section in the perldl.conf
in case someone looks for the module.  The plan is to keep it
there for a couple of releases just in case.  NOTE: it should
be straightforward to make a CPAN module of the Karma interface
if it is needed again...