Fix inconsistent/missing =for sig in documentation
I noticed from a recent mailing list post that the clip()
routine was missing a signature definition.  While I was
updating that, it was apparent that there were a number
of routines missing the =for sig to mark the Signature
information for the PDL::Doc to use.  This fixes all the
ones that I found.