Do not install modules using PDL::Slatec if Slatec is disabled
If Slatec support is disabled or detected as not available,
PDL::Slatec modules will not be installed. Regardless the fact, other
modules using PDL::Slatec get installed, despite of being unusable.

This patch ommits such modules from installation if Slatec has been
explicitly disabled.

This does not cover the case when disabled by autodetection. This is
because Slatec's Makefile.PL which does the detection is run after
Makefile.PLs of affected reverse dependencies.

Signed-off-by: Petr Písař <ppisar@redhat.com>